+++ src/database/entities/index.ts
@@ -3,3 +3,4 @@ export { EventData } from './event.js';
 export { MessageData } from './message.js';
 export { UserData } from './user.js';
 export { VoteData } from './vote.js';
+export { MemberAnniversaryRoleData } from './member-anniversary-role.js';
